requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

tinydb>=4.8.0
browser-cookie3>=0.19.1
apscheduler>=3.10.4
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
jsonpath-ng>=1.7.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.2
lxml>=5.3.0

# Dev & Build tools
watchdog>=6.0.0
pyinstaller>=6.0.0
